Setup project for trapping rain water and wrote failang test based on example 1 from leetcode
