+++ src/test/java/com/lperthel/datastructures/ContainerWithMostWaterTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
 
 class ContainerWithMostWaterTest {
-private Solution solution = new Solution();
+private ContainerWithMostWater solution = new ContainerWithMostWater();
 	@Test
 	void test_GivenVariableHeightContainer_ExpectMaxArea() {
 		final int[] height = {1,8,6,2,5,4,8,3,7};
