+++ src/test/java/com/lperthel/datastructures/LinkedListCycleIITest.java
@@ -2,8 +2,6 @@ package com.lperthel.datastructures;
 
 import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.concurrent.TimeUnit;
