+++ src/test/java/com/lperthel/datastructures/PartiallyReverseLinkedListTest.java
@@ -1,8 +1,7 @@
 package com.lperthel.datastructures;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class PartiallyReverseLinkedListTest {
