+++ src/test/java/com/lperthel/datastructures/ReverseALinkedListTest.java
@@ -2,9 +2,7 @@ package com.lperthel.datastructures;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.fail;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class ReverseALinkedListTest {
